requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.1.0
Pillow>=9.0.0
django-crispy-forms>=1.14.0